// src/app/recipes/enums/recipe.action.types.ts
export enum RecipeActionTypes {
  FETCH_RECIPE_LIST = '[RECIPES]_FETCH_RECIPE_LIST',
  FETCH_RECIPE = '[RECIPES]_FETCH_RECIPE',
  SELECT_RECIPE = '[RECIPE]_SELECT',
  ADD = '[RECIPES]_ADD',
  ADD_SUCCESS = '[RECIPES]_ADD_SUCCESS',
  ADD_FAILED = '[RECIPES]_ADD_FAILED',
  UPDATE = '[RECIPES]_UPDATE',
  DELETE_SUCCESS = '[RECIPES]_DELETE_SUCCESS',
  DELETE_FAILED = '[RECIPES]_DELETE_FAILED'
}